requirements.txt
baidu-aip==2.2.13.0
certifi==2019.3.9
chardet==3.0.4
idna==2.8
Pillow==6.0.0
PySide2==5.12.3
pytesseract==0.2.6
requests==2.22.0
shiboken2==5.12.3
urllib3==1.25.3
